[feat] shift module into a template, added script/project
Description
Since I want the default for the project to use a simple __main__.py
file
The python module template has been shifted into the .additional folder
Added a README.md into the .additional folder
