+++ {{ cookiecutter.project_name }}/.additional/project_template/{{ cookiecutter.project_name.replace('-', '_') }}/example.py
